content-security-policy/to-port-from-blink/1.1/reflected-xss-allow-expected.txt

ALERT: XSS
Tests that 'Content-Security-Policy: reflected-xss allow;' disables the XSSAuditor. This test passes if an alert is generated, and the script is allowed.
